MP4: in-place write keeps the mdat where it is (the padding box has 8 bytes of header; gaps of 1..7 bytes are rebuilt), chunk offsets stay correct
Found by writing copies of real files (Pentatonix, PowerDirector) and comparing the chunk offsets with the position of the mdat: the library shifted the media by 8 bytes without adjusting the offsets.

+++ MP4_MainContainer.h
@@ -27,6 +27,6 @@ public:
 	~CMP4_MainContainer(void);
 	CMP4Atom* find(CAtlString atomID, int count = 1);
 	void save(FILE *stream);
-	void adjustPadding(u32 size);
+	void adjustPadding(__int64 size);
 	void checkMetaBox();
 };
